backend/requirements.txt
# Web Framework and Server
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
mangum>=0.17.0
python-multipart>=0.0.5

# Database
pymongo>=4.0.0

# AWS
boto3>=1.26.0

# PDF Processing
pypdf==4.1.0
reportlab>=3.6.12

# Text Processing
textblob>=0.17.1
nltk>=3.8.1
numpy>=1.23.0

# HTTP and API
requests>=2.28.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Environment and Configuration
python-dotenv>=0.21.0

# YAML Support
pyyaml>=6.0.0

# Optional but recommended
types-requests>=2.31.0
types-python-dateutil>=2.8.19
types-PyYAML>=6.0.12
lark-parser>=1.1.2

# Mock S3
moto>=4.1.12
